chore: Moved create/join channel prompts outside of ChannelList
